Fix some statics errors. Started tidying up formatting.

+++ src/org/bentokit/flywire/plugins/playlist/PlayListPanel.java
@@ -45,115 +45,115 @@ public class PlayListPanel extends JPanel implements PlayListListener, MouseList
 	
 	PlayList playlist;  //The underlying playlist.
 	
-    public JPanel list; //Our display of the same playlist.
-    JPanel viewPanel, bottomPanel;
-    JScrollPane scrollPane;
-    JButton clearAllButton;
-    JLabel digitalTimer;
-    final JCheckBox cueModeCheckBox;
-    final JCheckBox repeatCheckBox;
-    MediaPanelFactory factory;
-    Map<MediaItem,MediaPanel> items = new HashMap<MediaItem,MediaPanel>();
+	public JPanel list; //Our display of the same playlist.
+	JPanel viewPanel, bottomPanel;
+	JScrollPane scrollPane;
+	JButton clearAllButton;
+	JLabel digitalTimer;
+	final JCheckBox cueModeCheckBox;
+	final JCheckBox repeatCheckBox;
+	MediaPanelFactory factory;
+	Map<MediaItem,MediaPanel> items = new HashMap<MediaItem,MediaPanel>();
 
 	public PlayListPanel() {
-        list = new JPanel();
-        list.setLayout(new BoxLayout(list,BoxLayout.Y_AXIS));
+		list = new JPanel();
+		list.setLayout(new BoxLayout(list,BoxLayout.Y_AXIS));
 
-        viewPanel = new JPanel(new BorderLayout());
-        viewPanel.add(list,"North");
-        viewPanel.add(new JPanel(),"Center");
+		viewPanel = new JPanel(new BorderLayout());
+		viewPanel.add(list,"North");
+		viewPanel.add(new JPanel(),"Center");
 
-        //Build the buttonPanel
-        bottomPanel = new JPanel(new BorderLayout());
-        JPanel leftPanel = new JPanel(new FlowLayout());
+		//Build the buttonPanel
+		bottomPanel = new JPanel(new BorderLayout());
+		JPanel leftPanel = new JPanel(new FlowLayout());
 
-        clearAllButton = new JButton("Clear All");
-        clearAllButton.setFont(new Font("Sans Serif",Font.PLAIN,15));
-        clearAllButton.addMouseListener(this);
-        leftPanel.add(clearAllButton);
+		clearAllButton = new JButton("Clear All");
+		clearAllButton.setFont(new Font("Sans Serif",Font.PLAIN,15));
+		clearAllButton.addMouseListener(this);
+		leftPanel.add(clearAllButton);
 
-        cueModeCheckBox = new JCheckBox("Cue Mode");
-        cueModeCheckBox.setFont(new Font("Sans Serif",Font.PLAIN,15));
-        cueModeCheckBox.setSelected(false);
-        leftPanel.add(cueModeCheckBox);
+		cueModeCheckBox = new JCheckBox("Cue Mode");
+		cueModeCheckBox.setFont(new Font("Sans Serif",Font.PLAIN,15));
+		cueModeCheckBox.setSelected(false);
+		leftPanel.add(cueModeCheckBox);
 
-        repeatCheckBox = new JCheckBox("Repeat");
-        repeatCheckBox.setFont(new Font("Sans Serif",Font.PLAIN,15));
-        repeatCheckBox.setSelected(false);
-        leftPanel.add(repeatCheckBox);
+		repeatCheckBox = new JCheckBox("Repeat");
+		repeatCheckBox.setFont(new Font("Sans Serif",Font.PLAIN,15));
+		repeatCheckBox.setSelected(false);
+		leftPanel.add(repeatCheckBox);
 
-        digitalTimer = new JLabel("00:00:00");
-        digitalTimer.setFont(new Font("Sans Serif",Font.BOLD,40));
+		digitalTimer = new JLabel("00:00:00");
+		digitalTimer.setFont(new Font("Sans Serif",Font.BOLD,40));
 
-        bottomPanel.add(leftPanel,"West");
-        bottomPanel.add(digitalTimer,"East");
+		bottomPanel.add(leftPanel,"West");
+		bottomPanel.add(digitalTimer,"East");
 
-        scrollPane = new JScrollPane(viewPanel);
+		scrollPane = new JScrollPane(viewPanel);
 
-        setLayout(new BorderLayout());
-        setBorder(BorderFactory.createTitledBorder(BorderFactory.createEtchedBorder(),"Playlist",
-                       TitledBorder.DEFAULT_JUSTIFICATION, TitledBorder.DEFAULT_POSITION,
-                       new Font("Sans Serif",Font.BOLD,20)));
-        add(scrollPane,"Center");
-        add(bottomPanel,"North");
-    }
-    	
+		setLayout(new BorderLayout());
+		setBorder(BorderFactory.createTitledBorder(BorderFactory.createEtchedBorder(),"Playlist",
+					TitledBorder.DEFAULT_JUSTIFICATION, TitledBorder.DEFAULT_POSITION,
+					new Font("Sans Serif",Font.BOLD,20)));
+		add(scrollPane,"Center");
+		add(bottomPanel,"North");
+	}
+		
 	public PlayListPanel(PlayList playlist, MediaPanelFactory factory, KeyListener k) {
-        this();
-        setFactory(factory);
-        setPlaylist(playlist);
-        setKeyListener(k);		
+		this();
+		setFactory(factory);
+		setPlaylist(playlist);
+		setKeyListener(k);		
 	}
 
-    public void setFactory(MediaPanelFactory factory) {
-        this.factory = factory;
-    }
+	public void setFactory(MediaPanelFactory factory) {
+		this.factory = factory;
+	}
 
-    public void setPlaylist(PlayList playlist) {
+	public void setPlaylist(PlayList playlist) {
 		this.playlist = playlist;
 
-        this.playlist.addListener(this);
+		this.playlist.addListener(this);
 
-        final PlayList finalPlaylist = playlist;
+		final PlayList finalPlaylist = playlist;
 
-        cueModeCheckBox.addItemListener(new ItemListener() {
-            public void itemStateChanged(ItemEvent e) {
-                finalPlaylist.setCueMode((e.getStateChange() == ItemEvent.SELECTED));
-            }
-        });
-        repeatCheckBox.addItemListener(new ItemListener() {
-            public void itemStateChanged(ItemEvent e) {
-                finalPlaylist.setRepeatMode((e.getStateChange() == ItemEvent.SELECTED));
-            }
-        });
-    }
+		cueModeCheckBox.addItemListener(new ItemListener() {
+			public void itemStateChanged(ItemEvent e) {
+				finalPlaylist.setCueMode((e.getStateChange() == ItemEvent.SELECTED));
+			}
+		});
+		repeatCheckBox.addItemListener(new ItemListener() {
+			public void itemStateChanged(ItemEvent e) {
+				finalPlaylist.setRepeatMode((e.getStateChange() == ItemEvent.SELECTED));
+			}
+		});
+	}
 
-    public void setKeyListener(KeyListener k) {
-        addKeyListener(k);
-        viewPanel.addKeyListener(k);
-        list.addKeyListener(k);
-        bottomPanel.addKeyListener(k);
-        clearAllButton.addKeyListener(k);
-        cueModeCheckBox.addKeyListener(k);		
-        repeatCheckBox.addKeyListener(k);		
+	public void setKeyListener(KeyListener k) {
+		addKeyListener(k);
+		viewPanel.addKeyListener(k);
+		list.addKeyListener(k);
+		bottomPanel.addKeyListener(k);
+		clearAllButton.addKeyListener(k);
+		cueModeCheckBox.addKeyListener(k);
+		repeatCheckBox.addKeyListener(k);
 
-    }
+	}
 
-    ///////////////////////////////////////////////////////////////////////////
-    // MouseListener interface ////////////////////////////////////////////////
-    ///////////////////////////////////////////////////////////////////////////
+	///////////////////////////////////////////////////////////////////////////
+	// MouseListener interface ////////////////////////////////////////////////
+	///////////////////////////////////////////////////////////////////////////
 	
-    public void mouseClicked(MouseEvent e) { if (e.getSource() == clearAllButton) playlist.clearAll(); }
-    public void mouseEntered(MouseEvent e) { }
-    public void mouseExited(MouseEvent e) { }
-    public void mousePressed(MouseEvent e) { }
-    public void mouseReleased(MouseEvent e) { }
+	public void mouseClicked(MouseEvent e) { if (e.getSource() == clearAllButton) playlist.clearAll(); }
+	public void mouseEntered(MouseEvent e) { }
+	public void mouseExited(MouseEvent e) { }
+	public void mousePressed(MouseEvent e) { }
+	public void mouseReleased(MouseEvent e) { }
 
-    ///////////////////////////////////////////////////////////////////////////
-    // PlayListListener interface /////////////////////////////////////////////
-    ///////////////////////////////////////////////////////////////////////////
+	///////////////////////////////////////////////////////////////////////////
+	// PlayListListener interface /////////////////////////////////////////////
+	///////////////////////////////////////////////////////////////////////////
 
-    public void setCueMode(boolean cueMode) { this.cueModeCheckBox.setSelected(cueMode); }
+	public void setCueMode(boolean cueMode) { this.cueModeCheckBox.setSelected(cueMode); }
 
 	@Override
 	public void playable(boolean isReady) {	; }
@@ -161,36 +161,36 @@ public class PlayListPanel extends JPanel implements PlayListListener, MouseList
 	@Override
 	public void stoppable(boolean stoppable) { ; }
 
-    public void mediaAdded(MediaItem item) {
-        final MediaItem finalItem = item;
-        final PlayList finalPlaylist = this.playlist;
-        
-   		ErrorHandler.info("PlayListPanel received notification of added item "+item.getURI());
-        MediaPanel panel = factory.createPanel(item);
-        panel.addMouseListener(new MouseAdapter() {
-            public void mouseClicked(MouseEvent e) {
-                System.err.println("PlayListPanel MediaItem clicked.");
-                finalPlaylist.remove(finalItem);
-            }
-        });
-        this.items.put(item,panel);
-		this.list.add(panel);	
-        this.scrollPane.getViewport().revalidate();
-		ErrorHandler.info("Item added to PlayListPanel "+item.getURI());		
-    }
+	public void mediaAdded(MediaItem item) {
+		final MediaItem finalItem = item;
+		final PlayList finalPlaylist = this.playlist;
+		
+		ErrorHandler.info("PlayListPanel received notification of added item "+item.getURI());
+		MediaPanel panel = factory.createPanel(item);
+		panel.addMouseListener(new MouseAdapter() {
+			public void mouseClicked(MouseEvent e) {
+				System.err.println("PlayListPanel MediaItem clicked.");
+				finalPlaylist.remove(finalItem);
+			}
+		});
+		this.items.put(item,panel);
+		this.list.add(panel);
+		this.scrollPane.getViewport().revalidate();
+		ErrorHandler.info("Item added to PlayListPanel "+item.getURI());
+	}
 
-    public void mediaRemoved(MediaItem item) {
+	public void mediaRemoved(MediaItem item) {
 		ErrorHandler.info("PlayListPanel received notification of removed item "+item.getURI());
 		this.list.remove(this.items.get(item));	
-        this.scrollPane.getViewport().revalidate();
-        this.factory.removeChild(this.items.get(item));
+		this.scrollPane.getViewport().revalidate();
+		this.factory.removeChild(this.items.get(item));
 		item.removeListener(this.items.get(item));
-        this.items.remove(item);
-		ErrorHandler.info("Item removed from PlayListPanel "+item.getURI());		
-    }
+		this.items.remove(item);
+		ErrorHandler.info("Item removed from PlayListPanel "+item.getURI());
+	}
 
-    public void timesChanged(long total, long sofar) {
-        digitalTimer.setText(TimeNumberFormat.formatTimeString(total-sofar));
-    }
+	public void timesChanged(long total, long sofar) {
+		digitalTimer.setText(TimeNumberFormat.formatTimeString((long)total-sofar));
+	}
 
 }
